Refactor rename process function to expandMacros
